back to the old last round assignement method post waitroom that works...
