+++ crosstalk/tests.py
@@ -6,8 +6,8 @@ from .models import Constants
 
 class PlayerBot(Bot):
     def play_round(self):
-        if self.round_number == 1:
-            yield pages.SetLastRound
+        # if self.round_number == 1:
+        #     yield pages.SetLastRound
         if self.round_number <= self.participant.vars['last_round']:
             yield pages.Decision, {"decision_high": 2, "decision_low": 4}
             # assert 'Both of you chose to Cooperate' in self.html  # no clue what this is
